feat(ui): simplify canvas component & hook API
