+++ invokeai/frontend/web/src/features/controlLayers/components/CanvasTabContent.tsx
@@ -1,18 +1,25 @@
-/* eslint-disable i18next/no-literal-string */
 import { Flex } from '@invoke-ai/ui-library';
+import { useAppSelector } from 'app/store/storeHooks';
 import { useScopeOnFocus } from 'common/hooks/interactionScopes';
 import { CanvasDropArea } from 'features/controlLayers/components/CanvasDropArea';
 import { Filter } from 'features/controlLayers/components/Filters/Filter';
-import { StageComponent } from 'features/controlLayers/components/StageComponent';
+import { CanvasHUD } from 'features/controlLayers/components/HUD/CanvasHUD';
+import { CanvasSelectedEntityStatusAlert } from 'features/controlLayers/components/HUD/CanvasSelectedEntityStatusAlert';
+import { InvokeCanvasComponent } from 'features/controlLayers/components/InvokeCanvasComponent';
 import { StagingAreaIsStagingGate } from 'features/controlLayers/components/StagingArea/StagingAreaIsStagingGate';
 import { StagingAreaToolbar } from 'features/controlLayers/components/StagingArea/StagingAreaToolbar';
 import { CanvasToolbar } from 'features/controlLayers/components/Toolbar/CanvasToolbar';
 import { Transform } from 'features/controlLayers/components/Transform/Transform';
 import { CanvasManagerProviderGate } from 'features/controlLayers/contexts/CanvasManagerProviderGate';
+import { TRANSPARENCY_CHECKERBOARD_PATTERN_DATAURL } from 'features/controlLayers/konva/patterns/transparency-checkerboard-pattern';
+import { selectDynamicGrid, selectShowHUD } from 'features/controlLayers/store/canvasSettingsSlice';
 import { memo, useRef } from 'react';
 
 export const CanvasTabContent = memo(() => {
   const ref = useRef<HTMLDivElement>(null);
+  const dynamicGrid = useAppSelector(selectDynamicGrid);
+  const showHUD = useAppSelector(selectShowHUD);
+
   useScopeOnFocus('canvas', ref);
 
   return (
@@ -29,7 +36,31 @@ export const CanvasTabContent = memo(() => {
       justifyContent="center"
     >
       <CanvasToolbar />
-      <StageComponent />
+      <Flex position="relative" w="full" h="full" bg={dynamicGrid ? 'base.850' : 'base.900'} borderRadius="base">
+        {!dynamicGrid && (
+          <Flex
+            position="absolute"
+            borderRadius="base"
+            bgImage={TRANSPARENCY_CHECKERBOARD_PATTERN_DATAURL}
+            top={0}
+            right={0}
+            bottom={0}
+            left={0}
+            opacity={0.1}
+          />
+        )}
+        <InvokeCanvasComponent />
+        <CanvasManagerProviderGate>
+          {showHUD && (
+            <Flex position="absolute" top={1} insetInlineStart={1} pointerEvents="none">
+              <CanvasHUD />
+            </Flex>
+          )}
+          <Flex position="absolute" top={1} insetInlineEnd={1} pointerEvents="none">
+            <CanvasSelectedEntityStatusAlert />
+          </Flex>
+        </CanvasManagerProviderGate>
+      </Flex>
       <Flex position="absolute" bottom={4} gap={2} align="center" justify="center">
         <CanvasManagerProviderGate>
           <StagingAreaIsStagingGate>
