+++ invokeai/frontend/web/src/features/controlLayers/hooks/useInvokeCanvas.ts
@@ -1,0 +1,51 @@
+import { useStore } from '@nanostores/react';
+import { $socket } from 'app/hooks/useSocketIO';
+import { logger } from 'app/logging/logger';
+import { useAppStore } from 'app/store/nanostores/store';
+import { useAssertSingleton } from 'common/hooks/useAssertSingleton';
+import { CanvasManager } from 'features/controlLayers/konva/CanvasManager';
+import Konva from 'konva';
+import { useLayoutEffect, useState } from 'react';
+import { useDevicePixelRatio } from 'use-device-pixel-ratio';
+
+const log = logger('canvas');
+
+// This will log warnings when layers > 5
+Konva.showWarnings = import.meta.env.MODE === 'development';
+
+const useKonvaPixelRatioWatcher = () => {
+  useAssertSingleton('useKonvaPixelRatioWatcher');
+
+  const dpr = useDevicePixelRatio({ round: false });
+
+  useLayoutEffect(() => {
+    Konva.pixelRatio = dpr;
+  }, [dpr]);
+};
+
+export const useInvokeCanvas = (): ((el: HTMLDivElement | null) => void) => {
+  useKonvaPixelRatioWatcher();
+  const store = useAppStore();
+  const socket = useStore($socket);
+  const [container, containerRef] = useState<HTMLDivElement | null>(null);
+
+  useLayoutEffect(() => {
+    log.debug('Initializing renderer');
+    if (!container) {
+      // Nothing to clean up
+      log.debug('No stage container, skipping initialization');
+      return () => {};
+    }
+
+    if (!socket) {
+      log.debug('Socket not connected, skipping initialization');
+      return () => {};
+    }
+
+    const manager = new CanvasManager(container, store, socket);
+    manager.initialize();
+    return manager.destroy;
+  }, [container, socket, store]);
+
+  return containerRef;
+};
